platform/requirements.txt
Flask==2.2.3
Flask_WTF==1.0.1
Flask_Login==0.6.2
Flask_SQLAlchemy==3.0.3
Flask_Migrate==4.0.7
Pillow==10.4.0
PyYAML==6.0.2
SQLAlchemy==2.0.37
Werkzeug==2.2.3
WTForms==2.3.3
gunicorn==20.1.0
email_validator==2.1.0
pytz
flask-paginate # 添加这一行
PyPDF2
Flask-SocketIO==5.3.4
gevent==23.9.1
gevent-websocket==0.10.1